backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
python-dotenv==0.19.0
openai>=1.0.0
opencv-python>=4.0.0
numpy>=1.19.0 